Move Toggle and Tag to dedicated directories
