+++ components/Destination.js
@@ -3,7 +3,7 @@ import { useState } from "react"
 
 import Button from "@components/Button/Button/Button"
 import Card from "@components/Card/Card"
-import Toggle from "@components/Toggle"
+import Toggle from "@components/Toggle/Toggle"
 
 export default function Destination({
   label,
